+++ api/src/main/scala/dx/api/DxFile.scala
@@ -1,6 +1,5 @@
 package dx.api
 
-import dx.AppInternalException
 import dx.api.DxPath.DxScheme
 import spray.json._
 import dx.util.{Enum, FileUtils}
@@ -42,6 +41,7 @@ case class DxFileDescribe(project: String,
                           size: Long,
                           state: DxState.DxState,
                           archivalState: DxArchivalState.DxArchivalState,
+                          tags: Option[Set[String]],
                           properties: Option[Map[String, String]],
                           details: Option[JsValue],
                           parts: Option[Map[Int, DxFilePart]])
@@ -50,6 +50,7 @@ case class DxFileDescribe(project: String,
     // all default fields must be specified to create this object,
     // so we only need to check non-default fields
     fields.diff(DxFileDescribe.DefaultFields).forall {
+      case Field.Tags       => tags.isDefined
       case Field.Properties => properties.isDefined
       case Field.Details    => details.isDefined
       case Field.Parts      => parts.isDefined
@@ -184,10 +185,10 @@ object DxFile {
               DxArchivalState.withNameIgnoreCase(archivalState),
               None,
               None,
+              None,
               None
           )
-        case _ =>
-          throw new Exception(s"Malformed JSON ${descJs}")
+        case _ => throw new Exception(s"Malformed JSON ${descJs}")
       }
 
     // populate the size field. It is missing from files that are in the open or closing states.
@@ -199,10 +200,11 @@ object DxFile {
 
     // optional fields
     val details = descJs.fields.get("details")
-    val props = descJs.fields.get("properties").map(DxObject.parseJsonProperties)
+    val tags = descJs.fields.get("tags").map(DxObject.parseJsonTags)
+    val properties = descJs.fields.get("properties").map(DxObject.parseJsonProperties)
     val parts = descJs.fields.get("parts").map(DxFile.parseFileParts)
 
-    desc.copy(size = size, details = details, properties = props, parts = parts)
+    desc.copy(size = size, details = details, tags = tags, properties = properties, parts = parts)
   }
 
   // Parse the parts from a description of a file
@@ -253,37 +255,9 @@ object DxFile {
   //   DxUtils.DxLinkKey: "file-F0J6JbQ0ZvgVz1J9q5qKfkqP"
   // }
   def fromJson(dxApi: DxApi, jsValue: JsValue): DxFile = {
-    val innerObj = jsValue match {
-      case JsObject(fields) if fields.contains(DxUtils.DxLinkKey) =>
-        fields(DxUtils.DxLinkKey)
-      case _ =>
-        throw new AppInternalException(
-            s"An object with key '$$dnanexus_link' is expected, not $jsValue"
-        )
-    }
-
-    val (fid, projId): (String, Option[String]) = innerObj match {
-      case JsString(fid) =>
-        // We just have a file-id
-        (fid, None)
-      case JsObject(linkFields) =>
-        // file-id and project-id
-        val fid =
-          linkFields.get("id") match {
-            case Some(JsString(s)) => s
-            case _                 => throw new AppInternalException(s"No file ID found in $jsValue")
-          }
-        linkFields.get("project") match {
-          case Some(JsString(pid: String)) => (fid, Some(pid))
-          case _                           => (fid, None)
-        }
-      case _ =>
-        throw new AppInternalException(s"Could not parse a dxlink from $innerObj")
-    }
-
-    projId match {
-      case None      => DxFile(fid, None)(dxApi)
-      case Some(pid) => DxFile(fid, Some(DxProject(pid)(dxApi)))(dxApi)
+    dxApi.dataObjectFromJson(jsValue) match {
+      case dxFile: DxFile => dxFile
+      case other          => throw new Exception(s"Not a file object ${other}")
     }
   }